Move preTraverse to outputnode because it is not big enough to be in its own file.
